requirements.txt
botornot
docopt>=0.6.0
eventlet>=0.19.0
Flask>=0.11.1
gunicorn>=19.6
networkx
pandas
psycopg2
python-dateutil>=2.0
pytz
pyyaml
schema
scrapy>=1.3
simplejson
SQLAlchemy>=1.0
sqlparse
tabulate
tweepy
ruamel.yaml
